Added A Section Table in course-search.html
added a table in the courseSearch.html after you click the "Search for Course" button on the mainpage.
Put some logic in the controller to hide a table and present information on where to go if you just click on the course search page without clicking a button on the main page.