thymeleaf edits
was able to edit mainpage. To not display courses that were "not available"
This was for the core courses and the electives. and took away the button for electives that don't have any available classes. Once you click on the electives. Elective groups that don't have any available courses wont be displayed