Insert grades/ schedule
now with the repositories and service layers in place I refactored the insert grade function so you can actually add grades and it adds that course to the transcript table and it properly removes that course from the registration table.
theres alot of commented out code from the RegisteredSection.java class. Its obsolete along with the methods in MajorService and SQL caller classes.