Displays the courses that dont have any sections available.
displays the course id for the courses that are not offered in the term on the students main page. The search for sections button is not displayed if there is no sections available.
Also there is some exception handling for the preReq page when you click on the course for the information tab. (and that course is one of the courses that aren't offered and doesn't have any information.